@@ -1,5 +1,5 @@
-//go:build (go1.16 && baremetal) || (go1.16 && js) || (go1.16 && wasi) || (go1.16 && windows)
-// +build go1.16,baremetal go1.16,js go1.16,wasi go1.16,windows
+//go:build baremetal || js || wasi || windows
+// +build baremetal js wasi windows
 
 // Copyright 2009 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
